// src/main/webapp/app/entities/dto/offer-card.api.ts
export class OfferCardApi {
  machine: string = '';
  offerId: number = 0;
  offer: string = '';
  offerPicture: string = '';
  description: string = '';
  tags: string[] = [];
  rate: number = 0.0;
  numberRates: number = 0;
  minimalPrice: number = 0.0;
  profilePicture: string = '';
  profileName: string = '';
  city: string = '';
  isLiked: boolean = false;
}